+++ database/migrations/2025_10_05_062107_create_purchased_medicine_items_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->foreignId('purchased_medicine_id')->constrained('purchased_medicines')->cascadeOnDelete();
             $table->string('drug_name');
-            $table->decimal('quantity', 12, 2);
-            $table->decimal('unit_price', 12, 2);
-            $table->decimal('subtotal', 12, 2);
+            $table->integer('quantity');
+            $table->integer('unit_price');
+            $table->integer('subtotal');
             $table->timestamps();
         });
     }
